requirements.txt
asgiref==3.7.2
certifi==2023.11.17
charset-normalizer==3.3.2
crispy-bootstrap5==2023.10
defusedxml==0.7.1
diff-match-patch==20230430
Django==5.0.1
django-admin-interface==0.28.3
django-colorfield==0.11.0
django-crispy-forms==2.1
django-extensions==3.2.3
django-import-export==3.3.6
django-modeltranslation==0.18.11
django-registration-redux==2.13
django-rosetta==0.10.0
django-tinymce==3.6.1
django-versatileimagefield==3.0
et-xmlfile==1.1.0
idna==3.6
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.1.2
pillow==10.2.0
polib==1.2.0
psycopg2-binary==2.9.9
python-decouple==3.8
python-magic==0.4.27
python-slugify==8.0.1
PyYAML==6.0.1
requests==2.31.0
sqlparse==0.4.4
tablib==3.5.0
text-unidecode==1.3
typing_extensions==4.9.0
urllib3==2.1.0
xlrd==2.0.1
xlwt==1.3.0
